+++ Tests/AfterTraining/test_after_training.py
@@ -23,7 +23,7 @@ from azureml.core import Model, Run
 
 from InnerEye.Azure.azure_config import AzureConfig
 from InnerEye.Azure.azure_runner import RUN_RECOVERY_FILE
-from InnerEye.Azure.azure_util import MODEL_ID_KEY_NAME, download_run_outputs_by_prefix, \
+from InnerEye.Azure.azure_util import MODEL_ID_KEY_NAME, download_run_output_file, download_run_outputs_by_prefix, \
     get_comparison_baseline_paths, \
     is_running_on_azure_agent, to_azure_friendly_string
 from InnerEye.Common import common_util, fixed_paths, fixed_paths_for_tests
@@ -35,23 +35,26 @@ from InnerEye.Common.fixed_paths_for_tests import full_ml_test_data_path
 from InnerEye.Common.output_directories import OutputFolderForTests
 from InnerEye.Common.spawn_subprocess import spawn_and_monitor_subprocess
 from InnerEye.ML.common import DATASET_CSV_FILE_NAME, ModelExecutionMode
+from InnerEye.ML.configs.other.HelloContainer import HelloContainer
 from InnerEye.ML.configs.segmentation.BasicModel2Epochs import BasicModel2Epochs
 from InnerEye.ML.deep_learning_config import CHECKPOINT_FOLDER, ModelCategory
 from InnerEye.ML.model_inference_config import read_model_inference_config
 from InnerEye.ML.model_testing import THUMBNAILS_FOLDER
 from InnerEye.ML.reports.notebook_report import get_html_report_name
+from InnerEye.ML.run_ml import MLRunner
 from InnerEye.ML.runner import main
 from InnerEye.ML.utils.config_loader import ModelConfigLoader
 from InnerEye.ML.utils.image_util import get_unit_image_header
 from InnerEye.ML.utils.io_util import zip_random_dicom_series
+from InnerEye.ML.visualizers.plot_cross_validation import PlotCrossValidationConfig
 from InnerEye.Scripts import submit_for_inference
-from Tests.ML.util import assert_nifti_content, get_default_azure_config, get_nifti_shape
+from Tests.ML.util import assert_nifti_content, get_default_azure_config, get_default_workspace, get_nifti_shape
 
-FALLBACK_SINGLE_RUN = "refs_pull_498_merge:refs_pull_498_merge_1624292750_743430ab"
-FALLBACK_ENSEMBLE_RUN = "refs_pull_498_merge:HD_4bf4efc3-182a-4596-8f93-76f128418142"
-FALLBACK_2NODE_RUN = "refs_pull_498_merge:refs_pull_498_merge_1624292776_52b2f7e1"
-FALLBACK_CV_GLAUCOMA = "refs_pull_498_merge:HD_cefb6e59-3929-43aa-8fc8-821b9a062219"
-FALLBACK_HELLO_CONTAINER_RUN = "refs_pull_498_merge:refs_pull_498_merge_1624292748_45756bf8"
+FALLBACK_SINGLE_RUN = "refs_pull_545_merge:refs_pull_545_merge_1626538212_d2b07afd"
+FALLBACK_ENSEMBLE_RUN = "refs_pull_545_merge:HD_caea82ae-9603-48ba-8280-7d2bc6272411"
+FALLBACK_2NODE_RUN = "refs_pull_545_merge:refs_pull_545_merge_1626538178_9f3023b2"
+FALLBACK_CV_GLAUCOMA = "refs_pull_545_merge:HD_72ecc647-07c3-4353-a538-620346114ebd"
+FALLBACK_HELLO_CONTAINER_RUN = "refs_pull_545_merge:refs_pull_545_merge_1626538216_3eb92f09"
 
 
 def get_most_recent_run_id(fallback_run_id_for_local_execution: str = FALLBACK_SINGLE_RUN) -> str:
@@ -87,10 +90,10 @@ def get_most_recent_run(fallback_run_id_for_local_execution: str = FALLBACK_SING
     return get_default_azure_config().fetch_run(run_recovery_id=run_recovery_id)
 
 
-def get_most_recent_model(fallback_run_id_for_local_execution: str = FALLBACK_SINGLE_RUN) -> Model:
+def get_most_recent_model_id(fallback_run_id_for_local_execution: str = FALLBACK_SINGLE_RUN) -> str:
     """
     Gets the string name of the most recently executed AzureML run, extracts which model that run had registered,
-    and return the instantiated model object.
+    and return the model id.
     :param fallback_run_id_for_local_execution: A hardcoded AzureML run ID that is used when executing this code
     on a local box, outside of Azure build agents.
     """
@@ -101,7 +104,18 @@ def get_most_recent_model(fallback_run_id_for_local_execution: str = FALLBACK_SI
     tags = run.get_tags()
     model_id = tags.get(MODEL_ID_KEY_NAME, None)
     assert model_id, f"No model_id tag was found on run {most_recent_run}"
-    return Model(workspace=azure_config.get_workspace(), id=model_id)
+    return model_id
+
+
+def get_most_recent_model(fallback_run_id_for_local_execution: str = FALLBACK_SINGLE_RUN) -> Model:
+    """
+    Gets the string name of the most recently executed AzureML run, extracts which model that run had registered,
+    and return the instantiated model object.
+    :param fallback_run_id_for_local_execution: A hardcoded AzureML run ID that is used when executing this code
+    on a local box, outside of Azure build agents.
+    """
+    model_id = get_most_recent_model_id(fallback_run_id_for_local_execution=fallback_run_id_for_local_execution)
+    return Model(workspace=get_default_workspace(), id=model_id)
 
 
 def get_experiment_name_from_environment() -> str:
@@ -158,7 +172,6 @@ def test_registered_model_file_structure_and_instantiate(test_output_dirs: Outpu
     model_name = tags["model_name"]
     assert model_inference_config.model_name == model_name
     assert model_inference_config.model_configs_namespace.startswith("InnerEye.ML.configs.")
-    assert model_inference_config.model_configs_namespace.endswith(model_name)
     loader = ModelConfigLoader(model_configs_namespace=model_inference_config.model_configs_namespace)
     model_config = loader.create_model_config_from_name(model_name=model_inference_config.model_name)
     assert type(model_config).__name__ == model_inference_config.model_name
@@ -260,8 +273,8 @@ def test_expected_cv_files_segmentation() -> None:
     assert run is not None
     available_files = run.get_file_names()
     for split in ["0", "1"]:
-        for mode in [ModelExecutionMode.TEST, ModelExecutionMode.VAL]:
-            assert _check_presence_cross_val_metrics_file(split, mode, available_files)
+        assert _check_presence_cross_val_metrics_file(split, ModelExecutionMode.TEST, available_files)
+        assert not _check_presence_cross_val_metrics_file(split, ModelExecutionMode.VAL, available_files)
     # For ensemble we should have the test metrics only
     assert _check_presence_cross_val_metrics_file(ENSEMBLE_SPLIT_NAME, ModelExecutionMode.TEST, available_files)
     assert not _check_presence_cross_val_metrics_file(ENSEMBLE_SPLIT_NAME, ModelExecutionMode.VAL, available_files)
@@ -368,7 +381,6 @@ def test_training_2nodes(test_output_dirs: OutputFolderForTests) -> None:
 
 
 @pytest.mark.after_training_2node
-@pytest.mark.skip("Test times out for unknown reasons.")
 def test_recovery_on_2_nodes(test_output_dirs: OutputFolderForTests) -> None:
     args_list = ["--model", "BasicModel2EpochsMoreData",
                  "--azureml", "True",
@@ -434,3 +446,76 @@ def test_download_outputs_skipped(test_output_dirs: OutputFolderForTests) -> Non
     download_run_outputs_by_prefix(prefix, test_output_dirs.root_dir, run=run)
     all_files = list(test_output_dirs.root_dir.rglob("*"))
     assert len(all_files) == 0
+
+
+@pytest.mark.after_training_single_run
+def test_download_non_existing_file(test_output_dirs: OutputFolderForTests) -> None:
+    """
+    Trying to download a file that does not exist should raise an exception.
+    """
+    run = get_most_recent_run(fallback_run_id_for_local_execution=FALLBACK_SINGLE_RUN)
+    does_not_exist = Path("does_not_exist.csv")
+    with pytest.raises(ValueError) as ex:
+        download_run_output_file(blob_path=does_not_exist, destination=test_output_dirs.root_dir, run=run)
+    assert str(does_not_exist) in str(ex)
+    assert "Unable to download file" in str(ex)
+
+
+@pytest.mark.after_training_single_run
+def test_download_non_existing_file_in_crossval(test_output_dirs: OutputFolderForTests) -> None:
+    """
+    Downloading a non-existing file when trying to load cross validation results
+    should not raise an exception.
+    """
+    run = get_most_recent_run(fallback_run_id_for_local_execution=FALLBACK_SINGLE_RUN)
+    config = PlotCrossValidationConfig(run_recovery_id=None,
+                                       model_category=ModelCategory.Classification,
+                                       epoch=None,
+                                       should_validate=False)
+    config.outputs_directory = test_output_dirs.root_dir
+    does_not_exist = "does_not_exist.txt"
+    result = config.download_or_get_local_file(run,
+                                               blob_to_download=does_not_exist,
+                                               destination=test_output_dirs.root_dir)
+    assert result is None
+
+
+@pytest.mark.after_training_hello_container
+def test_model_inference_on_single_run(test_output_dirs: OutputFolderForTests) -> None:
+    falllback_run_id = FALLBACK_HELLO_CONTAINER_RUN
+
+    files_to_check = ["test_mse.txt", "test_mae.txt"]
+
+    training_run = get_most_recent_run(fallback_run_id_for_local_execution=falllback_run_id)
+    all_training_files = training_run.get_file_names()
+    for file in files_to_check:
+        assert f"outputs/{file}" in all_training_files, f"{file} is missing"
+    training_folder = test_output_dirs.root_dir / "training"
+    training_folder.mkdir()
+    training_files = [training_folder / file for file in files_to_check]
+    for file, download_path in zip(files_to_check, training_files):
+        training_run.download_file(f"outputs/{file}", output_file_path=str(download_path))
+
+    container = HelloContainer()
+    container.set_output_to(test_output_dirs.root_dir)
+    container.model_id = get_most_recent_model_id(fallback_run_id_for_local_execution=falllback_run_id)
+    azure_config = get_default_azure_config()
+    azure_config.train = False
+    ml_runner = MLRunner(container=container, azure_config=azure_config, project_root=test_output_dirs.root_dir)
+    ml_runner.setup()
+    ml_runner.start_logging_to_file()
+    ml_runner.run()
+
+    inference_files = [container.outputs_folder / file for file in files_to_check]
+    for inference_file in inference_files:
+        assert inference_file.exists(), f"{inference_file} is missing"
+
+    for training_file, inference_file in zip(training_files, inference_files):
+        training_lines = training_file.read_text().splitlines()
+        inference_lines = inference_file.read_text().splitlines()
+        # We expect all the files we are reading to have a single float value
+        assert len(training_lines) == 1
+        train_value = float(training_lines[0].strip())
+        assert len(inference_lines) == 1
+        inference_value = float(inference_lines[0].strip())
+        assert inference_value == pytest.approx(train_value, 1e-6)